+++ include/kinfo.h
@@ -1,23 +1,25 @@
 #ifndef INCLUDE_KINFO_H
 #define INCLUDE_KINFO_H
 #include "multiboot.h"
+#include "stddef.h"
+#include "stdint.h"
 
 struct kinfo {
     void *kernel_start_phys;
     void *kernel_start_virtual;
     void *kernel_end_phys;
     void *kernel_end_virtual;
-    unsigned int phys_mem_lower;
-    unsigned int phys_mem_upper;
+    size_t phys_mem_lower;
+    size_t phys_mem_upper;
     struct multiboot_tag_framebuffer tagfb;
     struct multiboot_tag_module mods[2];
     void *rsdp;
-    int acpi_ver;
-    int is_paging_enabled;
+    int32_t acpi_ver;
+    int32_t is_paging_enabled;
 };
 struct kinfo *get_kernel_info();
 // Get kernel and hardware info
 void kinfo_init(struct multiboot_tag *mb);
-void kinfo_set_phys_mem_lower(unsigned int phys_mem_lower);
-void kinfo_set_phys_mem_upper(unsigned int phys_mem_upper);
+void kinfo_set_phys_mem_lower(size_t phys_mem_lower);
+void kinfo_set_phys_mem_upper(size_t phys_mem_upper);
 #endif
