+++ include/timer.h
@@ -1,7 +1,10 @@
 #ifndef INCLUDE_TIMER_H
 #define INCLUDE_TIMER_H
-int timer_init();
+#include "stddef.h"
+#include "stdint.h"
+
+int32_t timer_init();
 // Get clock tick count from PIT.
-unsigned int getticks();
-void delay(unsigned int ms);
+size_t getticks();
+void delay(size_t ms);
 #endif
